moved quest participants to RecyclerView

+++ Habitica/src/com/habitrpg/android/habitica/ui/MainDrawerBuilder.java
@@ -104,7 +104,7 @@ public class MainDrawerBuilder {
                         BaseMainFragment fragment = null;
                         Class newActivityClass = null;
 
-                        switch (drawerItem.getIdentifier()) {
+                        switch ((int)drawerItem.getIdentifier()) {
                             case SIDEBAR_TASKS: {
                                 fragment = new TasksFragment();
                                 break;
